primerCommit
This project was developed for a school assignment. It is a compiler developed in Java, created in collaboration with a classmate, so I cannot take full credit for it. As my native language is Spanish, the entire project is written in Spanish. Additionally, I have extensive documentation, including documents and PDFs, that explain the process and my contributions to the project while studying the course, If you are interested in obtaining more information such as documentation, you can contact me at my email.